# Fix 232ms UI hang in call duration timer