### 🔗 Issue Links
*N/A - Performance improvement identified through profiling*

### 🎯 Goal
Fix a 232ms UI hang that occurs every second during active calls, caused by expensive Swift runtime protocol conformance checking when accessing `@Published` properties in a timer callback.

### 📝 Summary
- **Eliminated 232ms/second UI hang** by caching `startedAt` value to avoid `@Published` property access
- **Moved duration timer to background queue** using `DispatchSourceTimer` instead of `Foundation.Timer`
- **Removed unnecessary `@objc updateDuration()` method** by calculating duration inline
- **Zero impact on functionality** - duration updates work exactly the same, just more efficiently

### 🛠 Implementation
The issue was discovered through Instruments profiling, which showed `swift_dynamicCast` and protocol conformance checking taking 232ms when the duration timer accessed the `@Published var startedAt` property.

**Root cause:** Every second, the timer would access `startedAt` through the `@Published` property wrapper, triggering expensive Swift runtime checks for protocol conformances.

<img width="1294" alt="Screenshot 2025-06-06 at 11 25 33 AM" src="https://github.com/user-attachments/assets/9d89257c-a4a9-4f74-af97-ca9bb878f2e8" />


**Solution:**
1. **Cache the value**: Added `_cachedStartedAt` to store the date without going through `@Published`
2. **Background timer**: Switched from `Foundation.Timer` (main thread) to `DispatchSourceTimer` (background queue)
3. **Minimal main thread work**: Only the final `duration` update happens on main thread

This approach eliminates both the protocol conformance overhead AND keeps the timer work off the main thread.

### 🎨 Showcase
|  Before  |  After  |
| -------- | ------- |
| 232ms hang every second during calls | Smooth 60fps during calls |
| Timer blocks main thread | Timer runs on background queue |
| `@Published` access in hot path | Cached value access |

### 🧪 Manual Testing Notes
1. Start a video call
2. Verify the call duration timer updates every second
3. Monitor performance with Instruments - no more periodic hangs
4. Verify timer stops correctly when call ends
5. Test timer behavior during reconnection scenarios

### ☑️ Contributor Checklist
- [x] I have signed the [[Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform)](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [x] This change follows zero ⚠️ policy (required)
- [x] This change should receive manual QA
- [ ] Changelog is updated with client-facing changes
- [ ] New code is covered by unit tests
- [ ] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (tutorial, CMS)